Made changes from feedback
Fixed an issue where the Two of Clubs would always be the final card in a deck shuffled using the Riffle shuffle. Now, the program will randomly choose which half of the split deck is merged first. 

Changed some variables and methods to private as they were only used in their respective classes.

Signed-off-by: amb64 <123978335+amb64@users.noreply.github.com>